Performance size_t and is empty
If you want to use var to check size you should
using size_t. If you want to check size of list
you should use private variable to optimization
cost.

Partially-Implements: Exercise-2

+++ Exercise-2/headers/MyList.hpp
@@ -9,7 +9,7 @@
 template <typename T>
 struct Node
 {
-    Node() {}
+    Node(){}
 
     T value;
     std::shared_ptr<Node<T>> next_node;
@@ -21,7 +21,7 @@ class MyList
     public:
         MyList();
         virtual ~MyList() = default;
-        unsigned int getSize();
+        size_t getSize();
         utils::RESULT_CODE back_push(T object);
         utils::RESULT_CODE front_push(T object);
         utils::RESULT_CODE front_pop();
@@ -33,7 +33,7 @@ class MyList
         bool isEmpty();
 
     private:
-        unsigned int list_size;
+        size_t list_size;
         std::shared_ptr<Node<T>> head;
         std::shared_ptr<Node<T>> tail;
 };
